requirements.txt
openai>=1.3.5
jieba
Pillow
pandas